src/requirements.txt
absl-py==1.3.0
anyio==3.6.2
Appium-Python-Client==2.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
astroid==2.12.13
asttokens==2.2.1
async-generator==1.10
attrs==22.1.0
autobahn==23.1.2
Automat==22.10.0
Babel==2.11.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
comm==0.1.2
constantly==15.1.0
contourpy==1.0.6
cryptography==40.0.2
cycler==0.11.0
daphne==4.0.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
Django==4.2.1
entrypoints==0.4
exceptiongroup==1.0.1
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.51.1
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
incremental==22.10.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
isort==5.11.4
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.17.3
jupyter-client==7.4.8
jupyter-core==5.1.0
jupyter-events==0.5.0
jupyter-server==2.0.1
jupyter-server-terminals==0.4.2
jupyterlab==3.5.1
jupyterlab-pygments==0.2.2
jupyterlab-server==2.16.5
kiwisolver==1.4.4
lazy-object-proxy==1.8.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook-shim==0.2.2
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.6.0.66
outcome==1.2.0
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil-resolve-name==1.3.10
platformdirs==2.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pylint==2.15.9
Pympler==1.0.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.4
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
seaborn==0.12.1
selenium==4.6.0
Send2Trash==1.8.0
sentry-sdk==1.18.0
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.6.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
thop==0.1.1.post2209072238
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.4.0
-e git+https://github.com/killuhwhale/appium.git@2d6fb2f71249c0d6e3da80e51c9c179e40b36240#egg=ultralytics
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
xmldiff==2.4
zipp==3.11.0
zope.interface==6.0
